ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
httpx==0.27.0
pydantic==2.9.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
